backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Date parsing
python-dateutil==2.8.2

# Scheduling
apscheduler==3.10.4

# Storage (optional)
python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
